Replaced std::size_t by unsigned int in all template parameters
The reason is that Eclipse has often problems to resolve std::size_t, which makes it very annoying to work with the containers. In most cases it is the alignment parameter, where it makes no difference, since it is only used for compile time computations with integers from 1 to 8. In other cases, the former use of std::size_t was not quite correct (e.g. in the storage types). For the fixed containers it was okay, but it does not really matter which type is specified as the default value.