Modified pragma pack around data structure declarations.
Since alignment requirement reductions by pragma pack trumps the
alignment requirements specified by alignas etc., to keep flexibility,
it is important not to pack the containers by default. For this
reason #pragma pack(push, 16) ... #pragma pack(pop) is used as guard
around all data structure declarations. It is assumed, that there is no
type with higher alignment requirements than 16. If there is one, the
guards have to be adjusted accordingly.